--- src/rust/src/task/mod.rs
@@ -1,39 +1,0 @@
-use core::{
-    future::Future,
-    pin::Pin,
-    task::{Context, Poll}
-};
-use alloc::boxed::Box;
-pub mod simple_executor;
-pub mod executor;
-pub mod utils;
-
-pub struct Task {
-    id: TaskId, // new
-    future: Pin<Box<dyn Future<Output = ()>>>,
-}
-
-impl Task {
-    pub fn new(future: impl Future<Output = ()> + 'static) -> Task {
-        Task {
-            id: TaskId::new(), // new
-            future: Box::pin(future)
-        }
-    }
-
-    pub fn poll(&mut self, context: &mut Context) -> Poll<()> {
-        self.future.as_mut().poll(context)
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord)]
-struct TaskId(u64);
-
-use core::sync::atomic::{AtomicU64, Ordering};
-
-impl TaskId {
-    fn new() -> Self {
-        static NEXT_ID: AtomicU64 = AtomicU64::new(0);
-        TaskId(NEXT_ID.fetch_add(1, Ordering::Relaxed))
-    }
-}
